лр_7/project_/requirements.txt
python-socketio
    aiohttp